Refactor Mailchimp API into lib/mailchimp for better separation from the NL provider
