+++ packages/core/src/type/index.ts
@@ -15,6 +15,11 @@ export * from "./email-mailing-recipient";
 export * from "./export-type-id";
 export * from "./filter-handlers";
 export * from "./login-data";
+export * from "./mc-batch";
+export * from "./mc-member";
+export * from "./mc-operation-response";
+export * from "./mc-operation";
+export * from "./mc-status";
 export * from "./network-service-map";
 export * from "./network-service";
 export * from "./newsletter-contact";
@@ -33,4 +38,5 @@ export * from "./select-result";
 export * from "./tag-assignment";
 export * from "./taggable-entity";
 export * from "./update-contribution-result";
+export * from "./update-mc-member-response";
 export * from "./update-newsletter-contact";
