+++ packages/core/src/type/mc-member.ts
@@ -1,0 +1,12 @@
+import { MCStatus } from "./mc-status";
+
+export interface MCMember {
+  email_address: string;
+  status: MCStatus;
+  interests?: { [interest: string]: boolean };
+  merge_fields: Record<string, string>;
+  tags: { id: number; name: string }[];
+  timestamp_opt?: string;
+  timestamp_signup?: string;
+  last_changed: string;
+}
