+++ packages/core/src/type/mc-operation-response.ts
@@ -1,0 +1,5 @@
+export interface MCOperationResponse {
+  status_code: number;
+  response: string;
+  operation_id: string;
+}
