+++ packages/core/src/type/mc-status.ts
@@ -1,0 +1,1 @@
+export type MCStatus = "subscribed" | "unsubscribed" | "pending" | "cleaned";
